+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 -   Feature: expose `@Debounce()` decorator for throttling dispatch actions
 -   Feature: add extension API for NGXS Data plugin
 -   Feature: expose storage extension as plugin
+-   Feature: add fire init option to `@Persistence()` (storage extension)
 -   Feature: improved `@Persistence()` decorator for stability
 -   Fix: use deep equals when state update in storage from another browser tab
 -   Fix: now `@Computed()` fields are recalculation when store has changed